CPF tipo Long Long

+++ src/Locadora/Locadora.cpp
@@ -43,10 +43,10 @@ void Locadora::listar_clientes_por_cpf(){
 }
 
 // Controle de Locacao
-void Locadora::alugar_midias(long cpf, std::map<int, int> locacoes){
+void Locadora::alugar_midias(long long cpf, std::map<int, int> locacoes){
     this->central_midia_locacoes.alugar_midias(cpf, locacoes, this->central_de_clientes);
 }
-void Locadora::devolver_midias(long cpf){
+void Locadora::devolver_midias(long long cpf){
     this->central_midia_locacoes.devolver_midias(cpf, this->central_de_clientes);
 }
